# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MachineEmu Rust API",
    "description": "API v2 for instances, live devices, and ticketed streams",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v2/image-imports/vmmanager-base": {
      "post": {
        "tags": [],
        "summary": "Start a vmmanager-sh base image import job",
        "operationId": "start_vmmanager_base_import",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportVmmanagerBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Start a vmmanager-sh base image import job",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v2/image-imports/{id}": {
      "get": {
        "tags": [],
        "summary": "Get an image import job",
        "operationId": "get_image_import",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get an image import job",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v2/image-imports/{id}/events": {
      "get": {
        "tags": [],
        "summary": "Stream image import progress",
        "description": "SSE progress events use event, id, and JSON data fields. Reconnect with Last-Event-ID to replay retained progress.",
        "operationId": "stream_image_import_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image import event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Image import not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v2/images": {
      "post": {
        "tags": [],
        "summary": "Register an image",
        "operationId": "register_image",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterImage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Register an image",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v2/images/{id}": {
      "get": {
        "tags": [],
        "summary": "Get an image",
        "operationId": "get_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get an image",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [],
        "summary": "Replace an image manifest (JSON or YAML)",
        "operationId": "put_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterImage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replace an image manifest (JSON or YAML)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ImportVmmanagerBase": {
        "type": "object",
        "required": [
          "source",
          "image_id",
          "engine_track",
          "target"
        ],
        "properties": {
          "engine_track": {
            "type": "string"
          },
          "image_id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        }
      },
...
      "RegisterImage": {
        "type": "object",
        "required": [
          "image_id",
          "engine_track",
          "target",
          "disk_sha256"
        ],
        "properties": {
          "disk_sha256": {
            "type": "string"
          },
          "engine_track": {
            "type": "string"
          },
          "firmware_sha256": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_id": {
            "type": "string"
          },
          "supported_engine_tracks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "target": {
            "type": "string"
          },
          "tpm_state_sha256": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}